+++ src/main/java/com/b1n_ry/yigd/compat/InvModCompat.java
@@ -2,6 +2,7 @@ package com.b1n_ry.yigd.compat;
 
 import com.b1n_ry.yigd.compat.misc_compat_mods.CommonProtectionApiCompat;
 import com.b1n_ry.yigd.compat.misc_compat_mods.OrpheusCompat;
+import com.b1n_ry.yigd.compat.misc_compat_mods.RespawnObelisksCompat;
 import com.b1n_ry.yigd.config.YigdConfig;
 import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.nbt.NbtCompound;
@@ -35,6 +36,8 @@ public interface InvModCompat<T> {
             CommonProtectionApiCompat.init();
         if (loader.isModLoaded("orpheus"))
             OrpheusCompat.init();
+        if (compatConfig.enableRespawnObelisksCompat && loader.isModLoaded("respawnobelisks"))
+            RespawnObelisksCompat.init();
     }
 
     String getModName();