+++ src/main/java/com/b1n_ry/yigd/compat/misc_compat_mods/OrpheusCompat.java
@@ -13,7 +13,7 @@ public class OrpheusCompat {
             if (!orpheusLyrePower.equals("keep") && !orpheusLyrePower.equals("both")) return;
 
             if (inventory.containsAny(stack -> stack.isOf(ItemsRegistry.ORPHEUS_LYRE.get()), s -> true, i -> true)) {
-                inventory.handleItemPairs(stack -> true, s -> true, i -> true, pair -> pair.setRight(DropRule.KEEP));
+                inventory.handleItemPairs(mod -> true, (item, slot, pair) -> pair.setRight(DropRule.KEEP));
             }
         });
     }